refactor(cli): extract each command into its own command_*.py file
Closes #1158

cli.py now contains only the Typer app bootstrap and shared helpers
(console, debug_option, enable_debug_logging, _print_logs,
OrderedCommands, OrderedCommandsWithMigrationHints, version_callback,
and the root callback). Each command lives in a dedicated module;
cli.py late-imports them to trigger @app.command registration, and
explicitly mounts the import/export sub-apps via app.add_typer.

- datacontract/command_{init,lint,changelog,test,ci,publish,catalog,api}.py (new)
- datacontract/cli_import.py -> datacontract/command_import.py (rename)
- datacontract/cli_export.py -> datacontract/command_export.py (rename)

Side effect: top-level --help now lists the direct commands in
alphabetical order (api, catalog, changelog, ci, init, lint, publish,
test) because OrderedCommands preserves registration order and cli.py
imports the modules alphabetically.
